Integrating randomness beacon
This commit integrates the randomness beacon into chain.

It is WIP, there's a list of outstanding items after the test plan.

The randomness beacon works by assuming synchronous network, in other
words it is expected that (almost) all the honest block producers can
confirm the correctness of their encrypted shares during the allocated
time during the epoch.

The first 10% of the epoch is allocated for committing to the secret
shares, the last 10% is allocated for the aggregation, and the time in
between is allocated for challenges.

The aggregation currently happens one point at a time, by having chunks
publish their shares at particular index until the point at that index
is completely aggregated, and then move to the next index.

*Changes to approvals*
For two blocks A <- B in two different epochs we previously expected the
approvals in A on B to be from the block producers from epoch B. It
didn't make much sense (e.g. the producer of A can't include their own
approval), thus I changed it so that it's block producers of the epoch
of A who approve B.

*Changes to Doomslug:*
Some changes were necesary to Doomslug. The biggest change is that I
switched Doomslug from using stake to the number of seats. This is
needed to unify requirements for randomness and block creation
(otherwise a block that has more than 50% stake might still have less
than 50% of all the seats approving it).

Besides it, added storage of random reveals to Doomslug both for
received approvals (so called witness), and for the approvals that will
be sent by Doomslug.

*Changes to light client block computation*
With the DKG, an epoch switch can be delayed if the aggregation isn't
finished yet. The light client block computation was assuming that the
epoch switches immediately once some block beyond height `epoch_end - 3`
is produced. To account for it, I had to change the computation of the
light client block to push back the block from which the computation is
carried out till the block that would have been the epoch switch without
the delay related to the randomness beacon.

*Some semi-related fixes discovered by tests in this change:*
- If a block B builds on top of A, and a chunk in B was received while A
was in `blocks_with_missing_chunks`, when A finally gets accepted, it
unblocks B before the stored chunk gets processed, and thus the chunk
gets re-requested. I changed `request_chunks` to process all the stores
partial encoded chunks before requesting any new ones.
- We were using `head` epoch id when collecting approvals, not the epoch
id of the actual approval. It is completely wrong, and with the above
change to the approvals started actually causing problems. Fixes #2137.

Test plan
---------
1. `primitives/tests/randomness.rs` tests basic transitions of DKG
information in chunk blocks and headers, and its validation.
2. `client/tests/randomness.rs` contains tests that test integration
with the client. All tests have detailed comments at the beginning
 - `dkg_chunk_requests_and_responses` tests basic transitions and
    message passing, as well as reveals. The test has block production
    time sufficient to have no forks, and thus can test many invariants
    that would not hold in a less optimistic setup.
    Also tests random reveals, and makes sure that any subset of
    revealed shares results in the same randomness beacon output.

Outstanding items:
-----------------
1. If the beacon is disabled, it resorts to the previous hash. Should be
the output of a VDF (and the seed for the real beacon should be the
output of the VDF as well)
2. Aggregation might be slow today (it pulls all the shares each time it
creates a chunk DKG header), though my early profiling shows it is not
the biggest bottleneck.
3. Most of the randomness-related things are pretty slow. With 8 block
producers creating and distributing a chunk during the commit phase both
take on the order of 30ms, rand reveal takes 20ms.
4. While some code for challenges exist, the most important part
(sending them around) is not written.
5. It is assumed that the block producers have ED25519 keys, but it is
not enforced anywhere.
6. Significantly better coverage with fuzzy tests is necessary.

Generally search for occurrences of MOO in the codebase for
unfinished parts.