feat(chain): Removing Nightshade Finality Gadget if favor of modified Doomslug
The discussion:
https://github.com/nearprotocol/NEPs/issues/54

The associated spec PR:
https://github.com/nearprotocol/NEPs/pull/57

Removing Nightshade Finality Gadget and changing Doomslug to have
tolerance of 2/3. Also removing the concept of score, since Doomslug's
fork choice rule is just height.

Also while refactoring found that the conditions to issue endorsements /
skips in Doomslug were not precise, and somewhat redundant, so they were
also rewritten. It resulted in unifying `largest_endorsed_height` and
`largest_skipped_height` into a single field `largest_target_height`.

The scope of changes besides removing NFG:
1. The lightclient was entirely based on NFG. To keep the code compile I
replaced it tentatively with something that I think is close to what it
will look like with Doomslug, but the full spec (and likely a follow-up
PR) are a future work.
2. The epoch switch condition and pushing last final block back are both
removed as unnecessary with the new finality rule
3. Removed `is_endorsement` field from the approval. Any approval that
has height difference of exactly one is an endorsement now. Also removed
`reference_hash` since it was only used by NFG.
4. `Block::produce` now has enough information to compute last final and
last ds final blocks, simplifying its API.

There was also an issue in Doomslug that was handling skips and
endorsements in the same way, while they must be handled differently
(see https://github.com/nearprotocol/NEPs/pull/57). I fixed the issue,
and also adjusted the stress test we already have to catch it (see
below)

Test plan
---------
Changes in the old Doomslug tests:
1. `test_endorsements_and_skips_basic`: all the cases in which an
approval for the next height which was not an endorsement was produced now produce no approval, since without NFG there's no need for approvals for the very next height which are not endorsements. Otherwise the test is unchanged.
2. `test_doomslug_approvals` I left the structure of the test the same,
but designed new scenarios, since the change in the logic of when a
block is ready to be produced is pretty drastic.

Changed `test_fuzzy_doomslug_liveness_and_safety` to ocasionally produce
multiple blocks per height by two out of eight participants, and made
the check that ensures final blocks are not reverted to use hashes, not
heights. Collectively that catches the bug with skips and endorsements
handled the same way. It stalls if only hash is used to include
approvals in blocks, and catches safety violation if only height is
used.