feat(network): Ideal connection strategy.
Introduce ideal connections strategy to handle active connections.
It is an interval where the number of active connections should be
in a healthy peer on a network with enought members. This range,
gives some margin to all peer to accept new peers, and at the same
time the network is rebalanced on new peers joingin and on peers
leaving.

Test plan
=========
connect_on_full_network already pass.
Five peers connect among themselves, at most 4 conenctions allowed,
but not all edges are createds since ideal rang is [2, 3].
When new peer joins the network it is able to connect with at least
two peers and other existing peers adjust its active connections to
keep it on this interval.